Revert "Move master and slave to lib"
This reverts commit 2126db6649b4671ea3e8baaf3aae4b0f6e4a6b6c.
