+++ slave/slave.go
